Admin css ok

+++ front/src/components/Admin.tsx
@@ -34,26 +34,29 @@ function Admin() {
   const [admins, setAdmins] = useState<Admin[]>([]);
   const navigate = useNavigate();
 
-  const handleDelete = (id: number) => {
-    const token = localStorage.getItem('jwtToken');
-    if (token) {
-      fetch(`${API_BASE_URL}/api/admins/${id}`, {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      })
-        .then(response => {
-          if (response.ok) {
-            setAdmins(admins.filter(admin => admin.id_pers !== id));
-          } else {
-            // No alert for error
-          }
+  const handleDelete = (id: number, name: string, firstname: string) => {
+    const isConfirmed = window.confirm(`Are you sure you want to delete ${firstname} ${name}?`);
+    if (isConfirmed) {
+      const token = localStorage.getItem('jwtToken');
+      if (token) {
+        fetch(`${API_BASE_URL}/api/admins/${id}`, {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         })
-        .catch(error => {
-          console.error('Error deleting admin:', error);
-          // No alert for error
-        });
+          .then(response => {
+            if (response.ok) {
+              setAdmins(admins.filter(admin => admin.id_pers !== id));
+            } else {
+              // No alert for error
+            }
+          })
+          .catch(error => {
+            console.error('Error deleting admin:', error);
+            // No alert for error
+          });
+      }
     }
   };
 
@@ -173,59 +176,65 @@ function Admin() {
     }
   }, [admins]);
 
-  const handleDeleteTeam = (id: number) => {
-    const token = localStorage.getItem('jwtToken');
-    if (token) {
-      fetch(`${API_BASE_URL}/api/teams/${id}`, {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      })
-        .then(response => {
-          if (response.ok) {
-            setTeams(teams.filter(team => team.id_team !== id));
-          } else {
-            // No alert for error
-          }
+  const handleDeleteTeam = (id: number, teamName: string) => {
+    const isConfirmed = window.confirm(`Are you sure you want to delete the team "${teamName}"?`);
+    if (isConfirmed) {
+      const token = localStorage.getItem('jwtToken');
+      if (token) {
+        fetch(`${API_BASE_URL}/api/teams/${id}`, {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         })
-        .catch(error => {
-          console.error('Error deleting team:', error);
-          // No alert for error
-        });
+          .then(response => {
+            if (response.ok) {
+              setTeams(teams.filter(team => team.id_team !== id));
+            } else {
+              // No alert for error
+            }
+          })
+          .catch(error => {
+            console.error('Error deleting team:', error);
+            // No alert for error
+          });
+      }
     }
   };
 
-  const handleDeleteManager = (managerId: number, teamId: number) => {
-    const token = localStorage.getItem('jwtToken');
-    if (token) {
-      fetch(`${API_BASE_URL}/api/teams/${teamId}/managers/${managerId}`, {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      })
-        .then(response => {
-          if (response.ok) {
-            // Update the state to remove the manager from the team
-            setTeams(prevTeams =>
-              prevTeams.map(team =>
-                team.id_team === teamId
-                  ? {
-                      ...team,
-                      managers: team.managers?.filter(manager => manager.id_pers !== managerId) || [],
-                    }
-                  : team
-              )
-            );
-          } else {
-            // No alert for error
-          }
+  const handleDeleteManager = (managerId: number, teamId: number, managerName: string, managerFirstname: string) => {
+    const isConfirmed = window.confirm(`Are you sure you want to delete ${managerFirstname} ${managerName} from the team?`);
+    if (isConfirmed) {
+      const token = localStorage.getItem('jwtToken');
+      if (token) {
+        fetch(`${API_BASE_URL}/api/teams/${teamId}/managers/${managerId}`, {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         })
-        .catch(error => {
-          console.error('Error removing manager:', error);
-          // No alert for error
-        });
+          .then(response => {
+            if (response.ok) {
+              // Update the state to remove the manager from the team
+              setTeams(prevTeams =>
+                prevTeams.map(team =>
+                  team.id_team === teamId
+                    ? {
+                        ...team,
+                        managers: team.managers?.filter(manager => manager.id_pers !== managerId) || [],
+                      }
+                    : team
+                )
+              );
+            } else {
+              // No alert for error
+            }
+          })
+          .catch(error => {
+            console.error('Error removing manager:', error);
+            // No alert for error
+          });
+      }
     }
   };
 
@@ -329,162 +338,170 @@ function Admin() {
       <h1>Admin Page - User: {userId}</h1>
       <p>Welcome to the admin section. Here you can manage users, settings, and more.</p>
 
-      <h2>List of Admins</h2>
-      <center><table>
-        <thead>
-          <tr>
-            <th>ID</th>
-            <th>Name</th>
-            <th>Firstname</th>
-            <th>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {admins
-            .sort((a, b) => a.name.localeCompare(b.name))
-            .map(admin => (
-              <tr key={admin.id_pers}>
-                <td>{admin.id_pers}</td>
-                <td>{admin.name}</td>
-                <td>{admin.firstname}</td>
-                <td>
-                  <button onClick={() => handleDelete(admin.id_pers)}>Delete</button>
-                </td>
-              </tr>
-            ))}
-        </tbody>
-      </table>
-      </center>
-
-      <h2>Add a New Admin</h2>
-      <div>
-        <select
-          value={selectedUserId !== null ? selectedUserId : ''}
-          onChange={(e) => setSelectedUserId(Number(e.target.value))}
-        >
-          <option value="">Select a user</option>
-          {users
-            .sort((a, b) => a.name.localeCompare(b.name))
-            .map(user => (
-              <option key={user.id_pers} value={user.id_pers}>
-                {user.name} {user.firstname} (ID: {user.id_pers})
-              </option>
-            ))}
-        </select>
-        <button onClick={handleAddAdmin} disabled={selectedUserId === null}>
-          Add as Admin
-        </button>
-      </div>
-
-      <h2>List of Teams and Their Managers</h2>
-      <center><table>
-        <thead>
-        <tr>
-          <th>Team ID</th>
-          <th>Team Name</th>
-          <th>Managers</th>
-          <th>Actions</th>
-        </tr>
-        </thead>
-        <tbody>
-          {teams
-            .sort((a, b) => a.team.localeCompare(b.team))
-            .map(team => (
-              <tr key={team.id_team}>
-                <td>{team.id_team}</td>
-                <td>{team.team}</td>
-                <td>
-                {team.managers && team.managers.length > 0 ? (
-                  team.managers
+      <div className="admin-container">
+        <div className="admin-section">
+          <h2>Admin</h2>
+          <div className="admin-content">
+            <div className="admin-list">
+              <table>
+                <thead>
+                  <tr>
+                    <th>ID</th>
+                    <th>Name</th>
+                    <th>Firstname</th>
+                    <th>Actions</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {admins
                     .sort((a, b) => a.name.localeCompare(b.name))
-                    .map((manager: TeamManager) => (
-                      <div key={manager.id_pers} className="manager-container">
-                        <span>{manager.name} {manager.firstname} (ID: {manager.id_pers})</span>
-                        <button className="manager-delete-button" onClick={() => handleDeleteManager(manager.id_pers, manager.id_team)}>Delete</button>
-                      </div>
-                    ))
-                ) : (
-                  'No managers'
-                )}
-                <div className="manager-container">
-                <select
-                  value={selectedManagerIds[team.id_team] ?? ''}
-                  onChange={(e) => setSelectedManagerIds(prevState => ({
-                    ...prevState,
-                    [team.id_team]: e.target.value ? Number(e.target.value) : null
-                  }))}
-                >
-                  <option value="">Select a manager</option>
-                  {users
-                    .filter(user => !team.managers?.some(manager => manager.id_pers === user.id_pers))
-                    .sort((a, b) => a.name.localeCompare(b.name))
-                    .map(user => (
-                      <option key={user.id_pers} value={user.id_pers}>
-                        {user.name} {user.firstname} (ID: {user.id_pers})
-                      </option>
+                    .map(admin => (
+                      <tr key={admin.id_pers}>
+                        <td>{admin.id_pers}</td>
+                        <td>{admin.name}</td>
+                        <td>{admin.firstname}</td>
+                        <td>
+                          <button className="manager-delete-button" onClick={() => handleDelete(admin.id_pers, admin.name, admin.firstname)}>Delete</button>
+                        </td>
+                      </tr>
                     ))}
-                </select>
-                <button
-                  className="manager-add-button"
-                  onClick={() => {
-                    const managerId = selectedManagerIds[team.id_team];
-                    if (managerId !== null && managerId !== undefined) {
-                      handleAddManager(team.id_team, managerId);
-                    }
-                  }}
-                  disabled={selectedManagerIds[team.id_team] === null || selectedManagerIds[team.id_team] === undefined}
-                >
-                  Add Manager
-                </button>
-                </div>
-                </td>
-                <td>
-                  <button onClick={() => handleDeleteTeam(team.id_team)}>Delete</button>
-                </td>
-              </tr>
-            ))}
-        </tbody>
-      </table>
-      </center>
+                </tbody>
+              </table>
+            </div>
+            <div className="admin-add">
+              <select
+                value={selectedUserId !== null ? selectedUserId : ''}
+                onChange={(e) => setSelectedUserId(Number(e.target.value))}
+              >
+                <option value="">Select a user</option>
+                {users
+                  .sort((a, b) => a.name.localeCompare(b.name))
+                  .map(user => (
+                    <option key={user.id_pers} value={user.id_pers}>
+                      {user.name} {user.firstname} (ID: {user.id_pers})
+                    </option>
+                  ))}
+              </select>
+              <button onClick={handleAddAdmin} disabled={selectedUserId === null}>
+                Add as Admin
+              </button>
+            </div>
+          </div>
+        </div>
 
-      <h2>Add a New Team</h2>
-      <div>
-        <input
-          type="text"
-          placeholder="Team Name"
-          value={newTeamName}
-          onChange={(e) => setNewTeamName(e.target.value)}
-        />
-        <button
-          onClick={() => {
-            const token = localStorage.getItem('jwtToken');
-            if (token && newTeamName) {
-              fetch(`${API_BASE_URL}/api/teams`, {
-                method: 'POST',
-                headers: {
-                  'Content-Type': 'application/json',
-                  Authorization: `Bearer ${token}`,
-                },
-                body: JSON.stringify({ team: newTeamName }),
-              })
-                .then(response => {
-                  if (response.ok) {
-                    // Refresh the page after adding a team
-                    window.location.reload();
-                  } else {
-                    alert('Error adding team');
+        <div className="teams-section">
+          <h2>Teams</h2>
+          <div className="teams-content">
+            <div className="teams-list">
+              <table>
+                <thead>
+                <tr>
+                  <th>Team ID</th>
+                  <th>Team Name</th>
+                  <th>Managers</th>
+                  <th>Actions</th>
+                </tr>
+                </thead>
+                <tbody>
+                  {teams
+                    .sort((a, b) => a.team.localeCompare(b.team))
+                    .map(team => (
+                      <tr key={team.id_team}>
+                        <td>{team.id_team}</td>
+                        <td>{team.team}</td>
+                        <td>
+                        {team.managers && team.managers.length > 0 ? (
+                          team.managers
+                            .sort((a, b) => a.name.localeCompare(b.name))
+                            .map((manager: TeamManager) => (
+                              <div key={manager.id_pers} className="manager-container">
+                                <span>{manager.name} {manager.firstname} (ID: {manager.id_pers})</span>
+                                <button className="manager-delete-button" onClick={() => handleDeleteManager(manager.id_pers, manager.id_team, manager.name, manager.firstname)}>Delete</button>
+                              </div>
+                            ))
+                        ) : (
+                          'No managers'
+                        )}
+                        <div className="manager-container">
+                        <select
+                          value={selectedManagerIds[team.id_team] ?? ''}
+                          onChange={(e) => setSelectedManagerIds(prevState => ({
+                            ...prevState,
+                            [team.id_team]: e.target.value ? Number(e.target.value) : null
+                          }))}
+                        >
+                          <option value="">Select a manager</option>
+                          {users
+                            .filter(user => !team.managers?.some(manager => manager.id_pers === user.id_pers))
+                            .sort((a, b) => a.name.localeCompare(b.name))
+                            .map(user => (
+                              <option key={user.id_pers} value={user.id_pers}>
+                                {user.name} {user.firstname} (ID: {user.id_pers})
+                              </option>
+                            ))}
+                        </select>
+                        <button
+                          className="manager-add-button"
+                          onClick={() => {
+                            const managerId = selectedManagerIds[team.id_team];
+                            if (managerId !== null && managerId !== undefined) {
+                              handleAddManager(team.id_team, managerId);
+                            }
+                          }}
+                          disabled={selectedManagerIds[team.id_team] === null || selectedManagerIds[team.id_team] === undefined}
+                        >
+                          Add Manager
+                        </button>
+                        </div>
+                        </td>
+                        <td>
+                          <button className="manager-delete-button" onClick={() => handleDeleteTeam(team.id_team, team.team)}>Delete</button>
+                        </td>
+                      </tr>
+                    ))}
+                </tbody>
+              </table>
+            </div>
+            <div className="teams-add">
+              <input
+                type="text"
+                placeholder="Team Name"
+                value={newTeamName}
+                onChange={(e) => setNewTeamName(e.target.value)}
+              />
+              <button
+                onClick={() => {
+                  const token = localStorage.getItem('jwtToken');
+                  if (token && newTeamName) {
+                    fetch(`${API_BASE_URL}/api/teams`, {
+                      method: 'POST',
+                      headers: {
+                        'Content-Type': 'application/json',
+                        Authorization: `Bearer ${token}`,
+                      },
+                      body: JSON.stringify({ team: newTeamName }),
+                    })
+                      .then(response => {
+                        if (response.ok) {
+                          // Refresh the page after adding a team
+                          window.location.reload();
+                        } else {
+                          alert('Error adding team');
+                        }
+                      })
+                      .catch(error => {
+                        console.error('Error adding team:', error);
+                        alert('Error adding team');
+                      });
                   }
-                })
-                .catch(error => {
-                  console.error('Error adding team:', error);
-                  alert('Error adding team');
-                });
-            }
-          }}
-          disabled={!newTeamName}
-        >
-          Add Team
-        </button>
+                }}
+                disabled={!newTeamName}
+              >
+                Add Team
+              </button>
+            </div>
+          </div>
+        </div>
       </div>
     </div>
   );
